settings: optimize the use of settings
We only use class Settings as QObject for localization. It doesn't
make sense that there are multiple Settings objects created. So let
only MainWindow create it and provide a pointer to it to DMMControl
as well as SerialPortCtr. There are also too many get functions. So
use the settings protected with these three classes as friends and
access settings directly.