settings: merge config into settings
It became unmaintainable having config and settings separated. So
merge the config into settings. This also improves the XML config
parser to only set the config for known setting names. We also
introduce helpers for common stuff like init of a combo box. This
helps us unifying code. We check for limits within the combo boxes
now for example.