dmmcontrol: rework error flow and add display handling
Also here we can have less indention by changing the error flow.
When disabling the beeper, we can use the chance to handle the
display on/off setting there. But we have to enable it again.
This is also the chance to get rid of the double system error
check.