report: output values in float notation of precision 12
Very old Excel XP has a bug: It interprets values in e-notation the
wrong way if there are more than two digits after the decimal point.
It gets -1.5643E-06 and this results in -1.56E-02. Only values in
floating-point notation are interpreted correctly. So use that
format with the precision of 12. This way the resolution is 1 pV or
1 pA which better than the precision the DMM can provide.